Project restart
After an error in pi configuration resulted in the loss of work, pi was essentially factory reset in advance of new workflow. These files represent a resumed project where thingspeak implementation has been removed in favour of blynk which is currently being implemented. Additionally, user_detection now uses mac address as opposed to ip and both it and hue_implementation are now separated for a more modular approach to the project.