Firebase implementation
Firebase DB implemented in prototype version. Next stage is to trigger the DB push upon application end.
The DB will then be used in conjunction with a basic glitch web application

Additional changes:
- JSON has been renamed and is now passed additional information upon user exit of network
- Additional hue light function to ensure light is neutral upon close and open of application
- Additional comments in time_checks

At this point the offset feature directly influences the user notification via visuals(lamp and display colour) and blynk notifications. I had debated also altering the displayed ETA as per the original spec but this may provide unnecessary complication. Still deciding on this feature but as is, only the above are affected by the offset.