Site completed
Glitch site and firebase realtime DB integration completed. Had intended to allow user to change time and location variables on site but this is proving too time consuming. 
Site currently displays last 3 journeys in addition to links to resources, an about and a link to this repo. Small changes to python code are largely renaming of variables.